+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added component event `.once` modifiers.
 - Added normalized `style` bindings for strings, arrays, and objects.
 - Added dynamic slot names for `<slot :name>`, `<template v-slot:[name]>`, and `<template #[name]>`.
+- Added default values for slot scope destructuring.
 
 ## 1.0.15 - 2026-05-07